Move to session listing page, add spoofed user agent
Started getting unsupported browser messages when moving to the
session listing page so tried to spoof the user agent. Still showing
that the browser is unsupported and not listing the sessions.

Suspect that it needs JS so may need to change tact and swap to
using Selenium instead of Requests.